requirements.txt
# Core dependencies
meltano==3.8
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# API Clients
requests==2.31.0

# dbt
dbt-postgres==1.10.0
